AiGeneratorStyle/Backend/requirements.txt
# Backend/requirements.txt
# Compatible with Python 3.13

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# HTTP Clients
httpx==0.25.2
requests==2.31.0

# Image Processing (for face photo uploads)
Pillow==10.4.0  # Updated for Python 3.13 compatibility
numpy==1.26.2  # Updated for Python 3.13

# Environment Management
python-dotenv==1.0.0

# File handling
aiofiles==23.2.1

# Development Tools
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional: Only if you plan to use these features
# ================================================

# 3D Processing (if you want to process GLB files)
# trimesh==4.0.5
# scipy==1.11.4

# Computer Vision (if you want face detection)
# opencv-python==4.8.1.78
# mediapipe==0.10.8

# AWS S3 (if you want cloud storage)
# boto3==1.29.7

# Database (if you want persistent storage)
# sqlalchemy==2.0.23
# alembic==1.13.0

# Authentication (if you want user accounts)
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4
# pyjwt==2.8.0

# Code formatting
# black==23.11.0